+++ internal/texture/api.go
@@ -1,21 +1,21 @@
 // Package texture implements different types of textures.
 package texture
 
-import "github.com/flynn-nrg/izpi/internal/vec3"
+import "github.com/flynn-nrg/go-vfx/math32/vec3"
 
 // UV represents a UV pair.
 type UV struct {
-	U float64
-	V float64
+	U float32
+	V float32
 }
 
 // Texture represents a texture.
 type Texture interface {
 	// Value returns the color values at a given point.
-	Value(u float64, v float64, p *vec3.Vec3Impl) *vec3.Vec3Impl
+	Value(u float32, v float32, p *vec3.Vec3Impl) *vec3.Vec3Impl
 }
 
 // SpectralTexture represents a spectral texture.
 type SpectralTexture interface {
-	Value(u float64, v float64, lambda float64, p *vec3.Vec3Impl) float64
+	Value(u float32, v float32, lambda float32, p *vec3.Vec3Impl) float32
 }
